Check heap property in C++ BinHeap, add verifier
When asserts are enabled, the heap property is now checked in the
siftUp and siftDown methods. The check iterates over (parts of) the
heap and verifies that each child is greater equal to its parent.
This will take a long time and should possibly be hidden through an
additional flag.

The BinHeapVerifier type enables checking the internal integrity of
the heap for users.